Complete Phase 5: Streaming Utilities
🎯 **MAJOR ACHIEVEMENT**: Added comprehensive streaming API with stateful compression/decompression

✅ **COMPLETED FEATURES**:
- Stateful Compressor struct for incremental compression
- Stateful Decompressor struct for incremental decompression
- stream_reader() and stream_writer() convenience functions
- Builder patterns with chunk size and progress callbacks
- Async support placeholder (future implementation)

📋 **TECHNICAL IMPLEMENTATION**:
- Created src/stream/mod.rs with public streaming API
- Implemented Compressor<W: Write + Seek> with full Write trait
- Implemented Decompressor<R: Read> with full Read trait
- Added comprehensive documentation and usage examples
- Integrated with existing compression/decompression systems
- Added async feature flag to Cargo.toml

🏗️ **API STRUCTURE**:
- Simple API: compress_file(), decompress_file()
- Builder API: CompressionBuilder, DecompressionBuilder
- Archive API: ArchiveBuilder, ArchiveExtractor
- Streaming API: Compressor, Decompressor, stream_reader(), stream_writer()

🔄 **READY FOR PHASE 6**: Error handling improvements

This completes the three-tier API architecture (Simple + Builder + Streaming) making mismall suitable for any use case from basic file operations to complex streaming pipelines.

+++ src/huffman/decoder.rs
@@ -103,7 +103,7 @@ pub fn decode<R: Read + Seek, W: Write>(
         let mut decoded_bytes_count = 0;
 
         // Calculate total number of actual data bits
-        let total_data_bits = (header.payload_actual_size as u64 * 8) - header.padding_bits as u64;
+        let total_data_bits = (header.payload_actual_size * 8) - header.padding_bits as u64;
         let mut bits_read_count: u64 = 0; // New counter for bits read
 
         while decoded_bytes_count < header.original_size {
